requirements.txt
yt-dlp>=2024.1.0
pydub>=0.25.1
shazamio>=0.6.0
spotipy>=2.23.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
audioop-lts>=0.2.1;python_version>="3.13"

# Optional: For Telegram bot
python-telegram-bot>=21.0

soundcloud-v2>=1.6.0

# For Discogs scraping
playwright>=1.40.0
